+++ sdt_probe_offset.h
@@ -1,3 +1,3 @@
 
-size_t get_sdt_probe_offset(int fd, char *probe_name);
+long get_sdt_probe_offset(int fd, char *probe_name);
